+++ src/nr/types/struct/struct/__init__.py
@@ -19,10 +19,10 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 # IN THE SOFTWARE.
 
+from .. import get_type_mapper
 from ..core.interfaces import Location
 from .struct import CustomCollection, Struct
 from .fields import *
-from .globals import get_type_mapper
 from .plugins import StructType
 
 
@@ -37,7 +37,7 @@ def serialize(mapper, data, py_type_def, type_mapper=None, _stackdepth=0):
   if type_mapper is None:
     type_mapper = get_type_mapper(None, _stackdepth + 1)
   datatype = type_mapper.adapt(py_type_def)
-  return mapper.deserialize(Location(None, None, data, datatype))
+  return mapper.serialize(Location(None, None, data, datatype))
 
 
 __all__ = [